This revision contains numerous small changes improving bps simulation efficiency.  The file convect2.F contains numerous correlations used to calculate the internal convection coefficients from building surfaces. Each correlation is implemented alongside code reporting the name of correlation in use and its relevant parameters to ESP-r's trace facilities.
Prior to this revision, the write statements
assigning the correlation name and parameter values
were performed regardless if the trace facilities
were in use. The code then checked the value of the
'dotrace' variable, and sent the character buffers to
the screen when requested by the user. 

This revision merely prefixes each write statement
with a 'if (dotrace)' statement to ensure that they
are not performed when tracing is disabled. These
changes have no impact upon simulation results, nor
upon ESP-r functionality. However, preliminary
testing suggests they yield a 5-10% reduction in
simulation runtime.




Former-commit-id: 3df9ca69f674ddd29d1759f465eae675c0e56b6b