Update modules to allow simulation parameter lists to include a directive that simulations run at several timesteps per hour can save results once per hour to reduce file size and data recovery time. This used to only be supported in interactive sessions.  The common block spfldat is affected and this happens in a number of source files.
An additional number is recorded in the *sps line of the
configuration file. It is backwards compatable and the default
assumption is that results are saved at every simulation timestep.   

This has no impact on the CETC tester models predictions however
it might be useful in the case of test models which generate
large resutls files which need to be scanned as part of the
tester.pl run.  

And esruprj/e2eplus.F also includes a user dialog if a window has
been defined without being in a parent surface  (see also patches
made to subroutine suredgeadj in a separate commit).  

And esrures/ipvrep.F also includes a patch for the token
separator used in the summary section of an IPV report. This
only has an impact on users who generate IPV reports.  




Former-commit-id: 01fc03289661c095e7f174ba11d9d25ddf8ca29d