The flow component multi-sensor controler was rarely used and not particularly functional. The controller has been revised to support the sort of control which is associated with window openings - e.g. if the zone temperature is within a particular temperature range AND the temperature at a boundary node is within a different temperature range the allow the window to open.
The multi-sensor controller supports up to 4 sensors
and allows the specification of an AND or OR logic
between the sensors. Each sensor can sense a different
thing and a different location.  

The result is that flow controls and flow networks which
used to require two or three controls in sequence can
now be represented by a single component and a single
control law.  

Also relaxed control of bidirectional flow components
to allow opening to larger than nominal size. Noticed
this in tests of range based control of doors.  

The number of miscel data items associated with flow controls
is increased to cope with the AND OR logic and so the building?.h
variants have also been updated.

This controller has been tested in a number of models and
appears to be robust (once you get your head around the
posibilities).  An exemplar model will be added to the
repository later after it is cleaned up.




Former-commit-id: 7b50ac800a18f95dbea5c477841b92e9c74fe96e