bpfcom.F added a range check when attempting to delete a control loop that does not exist. edcondb.F include a report of EN410 G-value for optical properties (as part of CIBSE TM33 work). Also update the menu text to use ISO 6946 U value (to bring it in line with other reports).
Noticed that during interative use of prj the initial
request to create a new obstruction file seemed to fail
but it always worked on the 2nd invocaton. Traced the
logic and found one common block variable which was not
being set and have patched edobs.F

Noticed that when attempting to create a model variant
which already contained shading files it was not possible
to ask for a model variant for modified obstructions. Extended
the logic in prj.F so that new obstruction files can be added as
well as modifications made to existing obstruction descriptions.

In prjqa.F removed a number of unused common block variables.
Also included code for calculating admittance factors for
a construction. It works for many constructions, but further
checking is required to ensure calculations are correct.




Former-commit-id: 6e78ad140d3190d063517211f01b0770d42f5955