Enhance subroutine askmultizone to work with zone obstruction files so that these can be manipulated in the make-me-a-version facility.
Noticed when trying to define an optical switching of a window which
was also a controlled component in a flow network that when the optical
control labeled as 'time based' was used
a) prj did not writing out one data line of the optical control type
and sensor setpoint into the tmc file (which causes an error when 
reading the file in) and   
b) that the 'time based' control was not being recognised in solar.F
unless the sensor setpoint was less than -98.   

Edited econstr.F to test for the 'time based' optical control.  

Edited include/building*.h to increase the number of optical
control periods from 3 to 8.  

Edted solar.F to recognise an optical control index of -99 as
well as a sensor setpoint less than -98.  

Edited edcon.F to clarify selection of 'time based' optical
controller and to ensure that if/else tests include the optical
control index of -99.  

Tested this with a model and confirmed that both the flow
and solar were being adjusted at the same time.  




Former-commit-id: b546211804ea9e9b331ecf195e3c58034275cfac