Noticed several irritants in terms of understanding mass flow results output. Truncated node/component names in graphs and lists made it difficult to be sure of items and export of tabular data did not support all that many columns of data. Also some of the logic in mfoutp.F and mfget.F was not documented.
In esrures/mfget.F and mfoutp.F added a new string variable
to hold a 32 char description of a flow element such as
a connection. This supplements 20 char description which
is often so truncated that it is difficult to be clear
which entity is which when editing the selection list.  

In esrures/mfoutp.F expanded the length of the string buffer
used for tabular flow data export from 124 to 248 char
to allow for more columns of data.

Update to common block GRTOOL to include a new variable
ILFLAG which signals whether labels associated with flow
items are less truncated and written out to a single
rather than a multiple in in flow tabular listings. Also
add in additional source code comments.  

Changes pass syntax check and have been compiled on Linux
GNU 3.4, Solaris GNU 3.4 and F90, MacOSX GNU 3.3. Tester.pl
(Linux) scripts run against released developer_branch of 19 
July 2006. Interface was tested on several existing models
and results sets.   




Former-commit-id: d3ad0861ffbcd5729b5addf91932d03c7b46d7ac