When creating a CFD domain cell dimensions were only available in the geometry section and such information is of considerable interest when defining boundary conditions.  Several steps taken to present the user with cell location data as well as adding an option to overlay the zone geometry over the gridding in 3D as well as some options to display only a sub-set of the 3D gridding.
In esruprj/edcfd.F added synopsis of each subroutine
at the top of the file. Expanded the introductory help message
to the user. Added option to list
cell coordinates and gridding when defining boundary
cell links to zone surfaces.  

In esrudfs/cgd.F extended facility to view cfd grid
to allow for a subset of grid lines to be drawn
in the 3D view as well as the option to superimpose
the zone geometry onto the 3D grid.  The standard
3D grid is sometimes cluttered. Also the superimposition
of the zone geometry allows users to more quickly check
to see if the cells align with the surfaces in the zone.

Subroutine vgrid3d has additional parameters to instruct
it on what to draw.

These changes have no impact on the contents of the
CFD domain description and thus no impact on predictions.
And a tester.pl (linux) was run anyway and indicates that
there is no impact on the tester models.



Former-commit-id: ab650c4f10c97c22a117f814c244742b6145b63d