Code in esrucom/econtrol.F EVCNTRLAW extended to include verbose messages for most of the ideal zone control laws. This should help when checking models and improves the QA report.
In esruprj/bpfcontrl.F update range in egetwr call for surface with heatpipe
control law (was set to restrict it to one of the first 8 surfaces).

Syntax checks made to econtrol.F and bpfcontrl.F and included in the
commit. Compiled on Solaris F90 compiler, Linux GNU 3.4 compiler,
Mac OSX GNU 3.3 compiler. tester.pl linux script run and no
differences reported (except for one plant model which tends always
to be out-of-tollerance on this linux box).



Former-commit-id: 1bf7b5a37e700a27087c7f444f77704c42969b12