Restore Makefile in esrubps to be Linux/OSX version.
Noticed when working with optical control (e.g. switching
optical properties) within zone tmc files that some useful
information was not being included in the QA report and this
made it difficult to confirm that the intended alternative
optical controls had been defined.

Updated esruprj/prjqa.F to include both the standard and
alternative solar transmission values (this makes it rather
easier to check). 

What is currently lacking is the name of
the alternative optical property set which is being used - 
but this will require keeping track of the users actions
when defining the optical controls. Try to do this in
a later update.

Tested this on several models which included optical controls.
Also ran the linux tester.pl script. Compiled on Solaris F90
compiler, OSX GNU 3.3, Linux GNU 3.4.



Former-commit-id: 84eb9c58a5fff515e1e9d6257402d8b1f39012ec