Extended the logic of edisp248 to cope with long words (e.g. a file name with explicit path) within a block of text that might be close to or slightly beyond the requested width of display line-warp.
Found an email suggesting that some error messages generated
when reading real numbers from files are truncated and are
not particular clear.

Revised egetwr to write to a message buffer 248 char
instead of 124 char. Revised egetwr to support longer
messages (e.g. more of the file that caused the read
error is included in the message. Egetwr makes use
of the 'long phrase' capabilities of edisp248 especially
in packed long lists of numbers as in zone viewfactor
files.

Later will update egwtwi etc. to support longer message
buffers.



Former-commit-id: 7518036e9efd7e6cb80c0ccdaa7e2b8f90dd168f