An old observation on the esp-r list that in text mode the project manager does not display optical properties within the optical database item editing menu.  Tracked the bug down to a mis-setting of the variable ino.
An old observation on the esp-r list that in text
mode the project manager does not display categories of
materials when this menu is entered.  Tracked the bug
down to a mis-setting of the variable ico. 

Tested this on Solaris F90 and GNU3.4, MacOSX GNU 3.3, 
Linux GNU 3.4. No difference in tester.pl run




Former-commit-id: d72b1944cd682eb4340b8ddff09b5db553916a1d