In esrubld/adjb.F and esrubld/input.F added in an additional boundary type for CEN 13791 partition type.
In esrubps/input.F added in an additional
boundary type for CEN 13791 partition type. 

In esrucfg/cfg.F and esrucfg/edonecon.F additional dialog and options
added to support the CEN 13791 connection (index is 7).

In esrucom/egeometry.F and emkcfg.F and esru_misc.F esystem.F scsys.F 
added to support the CEN 13791 connection (index is 7).

In esrudbm/esystemg.F added in connection type index 7.

In esruprj/edgeo.F edtopol.F  additional dialog and options
added to support the CEN 13791 connection.

These changes are required for acreditation of ESP-r in
several jurisdictions. CEN 13791 has a specific definition
of adiabetic which insists that if X W of solar radiation
falls on one side of a partition wall the exact same solar
radiation falls on the other side. This of course is only
physically possible in the 3rd planet of the zargon nebula
which has two suns ;-) This facility should not be
used for normal consulting purposes.

This has been compiled on Solaris F90 and GNU 3.4 as well
as OSX GNU 3.3, Linux GNU 3.4. It is been run through the
tester.pl scripts (Linux) on multiple occassions and
has no impact on any of the test models.




Former-commit-id: d0e963744f1e357a5720c98c71c722576e36feba