When compairing the logic in esrucom/scsys.F and esrucom/esystem.F both of which scan model configuration files I noticed that a number of the new key words in the configuration file would not have been recognised and might have given an error when scanned in to the esruaco esrucnv esrue2r esrueco esrugrd esruish and esrumrt modules (each of which use scsys.F rather than esystem.F).
The missing keywords are: myclm, bcd, wnd, slr_half_hr, resh2,
end_my_sets, gtmp.

These key words are not found in most models as they are NRCan
specific concepts. But the tester.pl is not sensitive to the
use of any of the above modules (which might be something to
talk about).
Compiled on Solaris F90, Linux GNU 3.4, OSX GNU 3.3, syntax
checked and tester.pl run before and after.



Former-commit-id: 065487a4934f2d39b39c7315afb2350c77588a62