This commit corrects a minor bug in the code responsible for toggling between 'short' and 'descriptive' plant file formats (see r381). The subroutine newnet, which displays the plant network creation menu in prj, uses a flag 'bInitialized' to ensure that the plant network file format was always set to 'short' when newnet is first called. However, bInitialized was not correctly set to 'true' after the initialization was completed. As a result, the user's selected file format presisted as long as the newnet menu was displayed, but was reset to 'short' when the user exited and then re-entered the menu.
These changes ensure that the value of bInitialized is correctly set 
after the first call to newnet, and permit the user's selected file 
format to persist when he leaves and returns to the plant network 
creation menu.

It's noteworthy that the code will always reset the file format to 
'short' at the start of each prj session. It might be improved to save 
the desired file format in the plant network file, permitting the 
user's choice to persist between prj sessions, but the plant network 
file doesn't currently support storage of meta-data. 

 



Former-commit-id: c91122fde9be4fef3757abddd39b61dd687e95d6