Merged r447 r448 r449 and r450 of the development branch into Jon_Hand branch which brings it up to date with r450 of the development branch.  No files were actually changed during this merge (as the development branch was taking files from Jon_Hand branch.
To get the commit to work a minor patch has been
included so that something happens....

Change to esruplt/pcomp2.F a typo in an error message for
CMP34C that gave the wrong CMP number. Patch suggested
by Denny Bauer of ESRU. It has no impact on esp-r
performance.



Former-commit-id: 6d3fa9b9cd111c78373364a52548ad6820039274