At a training venue noticed that there was no reliable exit from the exemplar model selection at the point of browse / own-it. Revised the code so that the cancel option is termed 'continue' and this exits to the opening menu of prj and added an additional choice of 'select another exemplar' which re-displays the list of exemplar models.
Also changed the logic in esruprj/cadio.F so that if 
an exemplar is determined to be off-line (its configuration 
file did not exist in the specified location) the user is 
returned to the list of exemplar groups. Altered esruprj/prj.F
to act on status of rexmpl call.  

Syntax check undertaken.
Compiled X11 on Solaris F90.
Compiled X11 on Mac OSX GNU 3.3.
Compiled X11 on Linux GNU 3.4
Tester.pl run with no differences

This revision does not alter the ESP-r model or predictions.
It is an interface only modification related to selection
of models. 




Former-commit-id: f5bc55caf511b426bf0f26050eb6fabc925a10f5