Added an exemplar model cellular_flh which implements an abstract representation of a floor heating system as well as a chilled ceiling.
Model has been added to the exemplars file.

Both summer and winter simulations have been
run to test the model (explanitory images were
captured and are associated with the model).

A description of the model is included in the log file in the
configuration folder along with a QA report.




Former-commit-id: 2877d29974dd689237d69a31757ff4397ad8d588