A recent focus on make it easier for users to browse and own models in the native windows version found patterns where spaces in folders caused a number of issues in the standard code: a) calls to xcopy were failing b) the creation of commands for 'doit' calls were failing    because of spaces in folders which required double quotes    and sometimes because of a mix of forward and backwards    slashes. c) the logic in the library routine fdroot sometimes failed    when there were spaces in file names d) where addpath was called there were no checks that the    application was running on a Windows machine and checks    would be appropriate
Several steps where taken to resolve most of these issues
and because of the inter-related nature of the code changes
they constitute a limited number of commits. This is the
first.

A new source file esrucom/dossupport.F was added which
carries out tasks needed for xcopy to work as well as
processing file names to be used in doit commands.
A variant of cmdfiledos called imgfiledos which does not 
add the double quotes is available for preping image file
names for passing to GTK display code.
Note dbdoscopy will need to be modified if/when database 
file names grow past 72 characters.

Where addpath is called there is a test for Unix/Windows
and in the latter case the longtfile is given further
processing. Addpath is called in quite a few places and
the same pattern is used in each case.

In a few cases where doit commands were being made up
with short file names a longer text buffer was used.


In esruprj/cadio.F update logic setting up and carying out xcopy
command. Documentation was added. Most debug statments currently
commented out. Added code to remember the name of the 
configuration file implied in an exemplar file string. 
Also added code to detect when native windows
version is about to call newprb after getting an exemplar file
name and setting parameter to newprb so that further folder
checks are not done.

In esruprj/eddb.F update logic for each occurrance of a
operating system copy command and corrected one mis-leading
file name in a warning message. dbdoscopy moved from eddb.F
to dossupport.F

Addpath and longtfile was being used (e.g. invocation of 
ish and mrt) and updated the files prj.F edobs.F edmrt.F 
edzone.F context.F eddb.F edtopol.F prjqa.F and cadio.F so 
that same pattern is used for addpath calls. 

In esruprj/prj.F there were several places with / embedded
in the code - these were changed to local variable fs which
is set based on Unix or Windows test.

Found a compiler directive mis-typed esruprj/edcfg.F line ~3365
changed ifdef GDK to ifdef GTK.

Testing

- syntax check done for both the X11 and GTK esru_lib code
- compiled Solaris GCC 3.4 GTK version
- compiled Solaris gcc 3.4 X11 and GTK
- compiled OSX GCC 3.3 X11
- compiled cygwin GCC 3.4 X11
- compiled MINGW GTK
- ran tester.pl with no reported differences
- Testing on native windows box with spaces in users default
  path indicated that an own-it command works and simulations
  can be run ok on the copied model without having to exit
  prj and click on the new model cfg file.

- In each case the following tests were carried out:
  a) a model was browsed and a simulation run and results recovered,
  b) owned model, ran a simulation and results recovered,
  c) started up prj at the level of the newly owned model top folder
     and a simulaton run and results recovered
  d) startup up prj in the newly owned model cfg folder, simulation
     run and results recovered.
  e) res started up in the folder with results file to check that 
     model files were found.

- Distributed a test native windows version to ~dozen people
  9 of who reported back on ability to browse and own models.




Former-commit-id: 0a8714b4e9b2ff52065773486bbe21f7c5c272bf