This commit modifies the battery model to support an additional control scheme. Previously, the battery's rate of charge and discharge was managed through the BATTERY_STATE/...,batDemandP common variable. In NRCan's RE-H2 modelling work, batDemandP was managed by the custom (and non-portable) re-h2-controller model and no provision for using the battery model with other control schemes existed. This revision adds an additional input to the battery model (input #48) that specifies the methodology used to manage the model's rate of charge and discharge. Three control types are possible:
Control type 1: 

  The battery's rate of charge and discharge is managed through 
  the BATTERY_STATE/...,batDemandP global variable. Setting input 
  number 48 to '1' invokes this control type, and provides 
  compatibility with the existing re-h2-controller model.

Control type 2:
  
  The battery manages its rate of charge and discharge to balance 
  the node to which it's connected, and no external controller is 
  required (details below.)  Setting input number 48 to '2' invokes 
  this control type.

Control type 3:
 
  The battery's rate of charge and discharge is managed using the 
  as-yet-unimplemented power-only-component controller facility. 
  Setting input number 48 to '3' invokes this control type, and 
  causes the battery model to stop the simulation with an error 
  message.

In addition, these changes add functionality necessary to support 
control type 2. When control type 2 is active, the battery model sums 
the enodlgt/...,penodl and enodlgt/...,penodg common variables to 
determine the power balance on its connected node, and sets the 
requested rate of charge/discharge to this value. The battery model 
will determine if sufficient capacity is available to meet the 
charge/discharge, and apply the actual power flux to the connected 
node.

The battery model's inputs to not presently permit specification of 
the maximum and minimum state-of-charge. Therefore, control type 2 
assumes the battery can be charged and discharged from 0 to 100% of 
its capacity, which may not be realistic in some circumstances.

Control type 2's reliance on the penodl and penodg global variables to 
determine the net balance on its connected node also introduces a 
potential pitfall. bps characterizes each power-only component in the 
order specified in the electrical network input file, and appends 
the amount of power produced or consumed by each component to 
the penodl and penodg arrays. Therefore, the battery model can only 
accurately determine the balance on the connected node if it appears 
after any other power component models that are connected to the same 
node in the electrical network file. 

A more general implementation would be possible if a robust facility 
existed to determine the actual power balance on each node. Such a 
facility would also require changes to the PCU model (mscomp.db1 
database entry #20), which for practical reasons adds its electrical 
flux directly to the penodl and penodg arrays instead of the ppowoc 
and qpowoc arrays, as is the convention for power-only-component 
models.  Provided the required changes are made to the PCU model and 
the power-balance-calculation facility added, control type 2 could be 
transparently modified to use the new calculation facility.

Unlike the plant domain, the electrical network does not impose overly 
strict syntax checking on the electrical network file. As a 
consequence, it is possible to specify power-only-components with 
fewer (or greater) numbers of inputs than those described in the 
power-only-component database record. The battery model takes 
advantage of this quirk, assuming that models specified with only 47 
inputs are from legacy networks created prior to the introduction of 
different control types, and therefore intended for use with control 
type 1. In this way, the battery model implicitly supports existing 
networks build for use with the previous version of th model.

While this approach introduces ambiguity to the battery model 
specification, the risk of incorrectly specified models and confused 
users is deemed to be small. Until now, the battery model was unusable 
without the res-h2-control module, and the res-h2-control module has 
not been released in the general distribution of ESP-r or in the 
'beta' version on development_branch. Thus, only NRCan has created 
working networks using legacy versions of the battery model.

These changes do not appear to affect the simulation results obtained 
with the battery model. This revision was compared against r687 on 
branch Alex_Ferguson, and no changes were found. Admittedly, r687 did 
not bring across all revisions from the renewable_hydrogen branch, and 
some code in the battery model may not have been exercised. Further 
testing is required before these changes are incorporated into 
development_branch. 

Finally, this revision and r687 are intended to provide Christianne 
Aussant and Ali Syed of Ryerson University with a working battery 
model for their CHP and renewable energy studies. They should *NOT* be 
merged into development_branch. Instead, they will be merged on to the 
Ryerson branch, where all future development for control type 2 will 
occur. 





Former-commit-id: 201d0e9605015343a8f6849cb0a111bcab958df2