This commit modifies the power-conditioning model to support two way power flow between nodes. Previously, the power conditioning model required the user to define the 'sending node' and 'receiving node', and assumes the power will flow from the sending node node to the receiving node to i) meet the load on the receiving node, or ii) balance a surplus in generation on the sending node. This change adds a third configuration, in which the 'receiving node' is monitored by the PCU. When a surplus of power exists on the receiving node, the PCU transfers this amount to the sending node, and when a deficit exists, the PCU pulls this amount from the sending node.
These changes allow modelling of AC/DC networks in which electrical 
storage is coupled to the DC bus and generation devices are connected 
to the AC bus. This configuration can be specified by setting the 
first input ('Operating mode of PCU (-)') to '3'. Specification of the 
sending and receiving nodes remains the same, and as before, the PCU 
must appear after any other power-only components that are coupled to 
the same 'receiving' node. 

These changes do not appear to affect the simulation results obtained
with the battery model. This revision was compared against r687 on
branch Alex_Ferguson, and no changes were found. 

Finally, this revision and r687 are intended to provide Christianne 
Aussant and Ali Syed of Ryerson University with a working PCU model 
for their CHP and renewable energy studies. They should *NOT* be 
merged into development_branch. Instead, they will be merged on to the 
Ryerson branch.






Former-commit-id: fabb30dd92195c7d64b2ba1ebe2ca796000212b2