a similar change as the previous one (log message copied):
This "relaxes" the limits of a "FAILURE" message that appears when
users select the CEN fixed convective coefficients. The highest value
that the was allowed was 20 W/m2K and this was changed to 30 W/m2K.
This can be applied when using the CEN hc as a combined convect. and 
radiative coefficient (by setting surface emiss. to 0) and the user tries
to achieve the CEN external surface resistance value of 0.04 m2K/W. This
attempt requires a combined convect./radiative coefficient of 25 W/m2K 
(i.e. 1/0.04). For this reason the "FAILURE" message appears now for values
above 30 W/m2K.




Former-commit-id: 403ae00d2200470ba2d5e5dc4a8375e7f706d3df