This revision completes functionality allowing H3Kreports to optionally save time-step data into a scratch file, instead of pushing it onto the heap.  The file is formatted as a list-of-lists; for each variable reported by h3kreports, the file contains all time-row data. Data for subsequent variables are appended at the end of the file. For instance, a file containing N time row data from two variables would look like this:
 VARIABLE:  ___________ V 1 _____________   ____________ V 2 ___________
           /                             \ /                            \
 TIME ROW: | __Ts1__  __Ts2__     __TsN__ | __Ts1__  __Ts2__     __TsN__ |
           |/       \/       \   /       \|/       \/       \   /       \|
 FILE DATA:|---8B---|---8B---|...|---8B---|---8B---|---8B---|...|---8B---|
 
There are no delimiters in the file; each piece of data is stored as
a double (~8 Bytes on most systems)

The facility does not need to know the total number of variables that
will be reported---it merely appends new variables on to the end of the
current file. But it must know the total number of timesteps expected
in a simulaton, which prevents its use with variable timestep.

These changes significantly reduce bps's memory requirements when run 
with h3kreports during long simulations at short time steps. The 
facility can be activated by including the <save_to_disk>true</save_to_disk> 
tag in the input.xml file. 




Former-commit-id: d5558e0537590b06d2659cd4ddf94569ae8b6abd