This commit makes numerous small improvements to the IMS model. Most importantly, it implements a coupling between the IMS unit's casual gains and the casual gains imposed in the space. The IMS model now computes additional heat exchange that is unaccounted for in the ideal control loop (which may include ventilation loads, gains associated from electrical use and skin losses), and these are appended to the zone temperature state equations.
In the IMS model, the casual gain coupling also does double-duty. Even 
though ESP-r's idealized controls can be configured to limit the 
amount of heat injection to match an HVAC system's capacity, the IMS 
model may determine that not all of the unit's capacity can be 
delivered. For instance, excessive ventilation loads or cocurrent 
water-heating loads may limit the available space heating capacity. In 
these conditions, the IMS model appends the resulting space heating 
deficit to the computed casual gains, and the undelivered heat will be 
accounted for during solution of the zone-side energy balance on the 
subsequent time step.





Former-commit-id: 18236c5a41d3def1d5bd0088f0ba0b9268db1bda