Commit adapts "building.h" to our standard sizes: 48 zones and 62 surfaces per zone. This commit is not aimed for the development_branch.
Comments to Commit 2295 (now that I know how it works :^) )
building.h: As above
cfd.h: Adapted MNSBZ to MS = 62 (Note: In the distribution there is a discrepancy between these variables), changed number of cells to 62 in each direction.
cfcalc.f: Changed "minimum number of iterations" for subsequent timesteps from 20 to 3.

Commit 2295 is not meant for the development_branch.



Former-commit-id: 6437a3f80de5b0b61fc931f88a21440403dba8ad