This commit replaces output of CFD results (time series) in TECPlot format by output in ParaView format. ParaView is Open Source (www.paraview.org).
At this stage, only temperature and velocity values are written to file. However, extension to the additional values available for TECPlot output can easily be done in the new function "Paraviewend()" which is essentially a clone of TECPlotend(). Both reside in file cfutil.f. TECPlotend() is not used due to the replacement of the call as opposed to an extension of calling possibilities in res.

Initial tests suggest that the output is as stable and correct as the TECPlot output.



Former-commit-id: 2a085d0f1e02d88dcd3897fad3aeba3a574210e9