Heat transfer coefficient for ventilated vertical channels (i.e. ventilated double facades) according to ISO 15099:2003 included.
If a flow network is defined (should be for such models), the mean channel velocity is taken from the flow network results.

Activation by .htc-file, control type 3, ICOR=17.

The >calculations< are checked for example cases and found O.K. The >model< for the described situation should be validated at some future point.



Former-commit-id: 43d445d9b5d820b84eaf3fbb3271a708edf53c62