Update the documentation associated with the INNO Setup installer used to create the Native Windows version of ESP-r. Also add several of the files included in the Native Windows Installer so these can be kept in sync with the evolution of ESP-r.
Users will notice that the dialogue has been updated in
the installer. Developers will now have more of the files
needed to create their own installer for Windows.

Revised the INNO iss script to remove the ESP-r cookbook
and the tool comparison pdf files from the installer as
these are available on the web.

No impact on predictions.




Former-commit-id: 38bba8d05a8d3f279c9564bd0cfc514ee55caecf