Simulation teams passing models between computers will often face the problem that a model includes an explicit path to standard corporate databases but when moved to another machine which has ESP-r installed to a different location the files do not resolve correctly and manual editing is required.
The idea is to alter the format of the configuration file
so the tags e.g. *opt has a variant *stdopt which says
the the name that follows is to be appended to the standard
location of databases. Thus
  *opt /usr/esru/esp-r/databases/optics.db2
becomes
  *stdopt optics.db2

The *stdopt signals the code to look at the folder
where ESP-r was compiled to. This does make the
assumption that the contents of the standard databases
are the same on the two machines. This MIGHT also
make it much easier for exemplar models to be installed
(less need for the Install script editing the paths
to the databases). This has yet to be proven.

Created a new subroutine findwhichdbpath which takes
a file name and determines if it is a local database,
a database with an explicit path or a database in the
standard installed database folder.

Extended the data within esprdbfile.h to hold an array
of indicies that remember the folder state for each
of the databases (but only optics is currently instanciated).

The initial implementation of findwhichdbpath is for
the optical database which is a 144 char file name. 
The pattern will next be extended to the profiles 
database (also a long string).

Work is needed to get the other database file names
lengthened from 72 to 144 characters. 

Users will probably not notice any change. 

Developers might notice there is an additional token in the
model configuration file and the path may seem to
be missing from the optical database if a standard
one is being used.  In the longer term the additional
logic and string buffer length should simplify working 
with databases and reduce the likelyhood of string 
buffer overflows.

Testing
- Compile OSX GCC 3.3 X11
- A sequence of debug sessions to arrive at a point where
  the user can work with an existing model, copy the default,
  optics, edit the existing optics, etc.
- This commit to allow a syntax check to be run.




Former-commit-id: f9af6a6a4825be28e60d4c63de63712ce51c152c