This commit introduces optical controls for bidirectional datasets in a similar way as for the other control domains. There is also some code for replicating the TMC controls to this structure but it requires significant changes in solar.F and it is left for future work (some code for this puropse has been commented and will exist in the branch). In detail:
sensor data options are similar as those for TMC controls:
      elseif(icfoc.eq.5)then
        WRITE(VERT(1),'(a)') 'a senses current zone db temp         '
        WRITE(VERT(2),'(a)') '  senses an ambient condition...      '
        WRITE(VERT(3),'(a)') 'b   dry bulb temperature              '
        WRITE(VERT(4),'(a)') 'c senses incident radiation at surface'
        WRITE(VERT(5),'(a)') 'd senses lux levels                   '
        WRITE(VERT(6),'(a)') '  ------------------------------------'

For lux levels, however, only the so called Lux level control is available and with some limitations (manual switch control is not
available). This is further explained in solar.F

The actuator will ask for which construction to control. This will be matched at a later stage with a users selection in the advanced
optics menu in order to use it for bidirectional datasets control.

Period of validity: 
as for the other controls.

Control laws: 
The first law is not active. This is meant to be optical properties switching law and the purpose is to have the same
functionality as for the TMC controls. The code in solar.F does not make it easy to implement this it at the moment. It appears on
the interface witha note (N/A) next to it and some background supporting code exists.

The second control law is for the bidirectional datasets. The user has the option to specify 6 setpoints and a relevant dataset for
each of these setpoints (i.e. 12 items, using an EMENU call to make it easy to display). 6 datasets are used because sample
bidirectional files seem to have maximum of 6 datasets. The code sorts the point in a descending order 
and if the sensed variable is above the specified setpoint it uses the relevant dataset. If the sensed value is betwen the 2nd and 3rd setpoint,
for example, it will use the dataset related to the 3rd setpoint. This control law is used in the solar.F where a new subroutine for
controlling solar and optics related items has now been added.

This commit is also to allow a syntax check.
It may also be useful to start moving this towards the development branch to reduce maintance time and allow other people to base
their developments on these controls too (e.g. Aizaz's work for day types).




Former-commit-id: c098575e8463f806cec683e3ae8eaae3903ccd69