Notes on manual conflation of source code from prj_dev branch (prior to Alexs overwriting of prj_dev with the develpment branch) with Jon_Hand branch.
Changes also include code in the prj_dev that Aizaz has been working 
on (3233, 3234 3235 3241 3242 3243 3244 3253 3256 and r3257). 

The Jon_Hand_3132 folder had been updated to revision 3245 prior to
this manual conflation taking place. 

This commit brings into Jon_Hand branch essentailly all of the backlog
of code that was pending in prj_dev.

tester.pl indicates no differences in predictions against the
develpment branch or against the prior verson of Jon_Hand branch.

Syntax checks indicate mostly a reduction in the number of
warnings and information messages as well as fewer errors.

Summary of changes within the code from prj_dev needs to 
be created.

This commit is to allow testing and compilation on other
computer types.

Testing
- Compile Linux GCC 3.4 X11
- Compile Linux GCC 3.4 GTK
- tester.plfor X11 run against develpment branch with no
  differences reported.
- Several iterations of syntax checks carried out.




Former-commit-id: 93d4573738a29f57e35b23c694860ace68d19706