This commit writes out the input.xml file based on description specified in the the so called "flat file". Example of the syntax needed for this: *start_xml building/all_zones/supplied_energy/heating building/all_zones/supplied_energy/cooling building/all_zones/envelope/all_components/heat_loss building/all_zones/envelope/all_components/heat_gain building/all_zones/envelope/infiltration/heat_gain building/all_zones/envelope/infiltration/heat_loss building/all_zones/internal_gains/total building/all_zones/insolation/total building/all_zones/insolation/adverse building/all_zones/insolation/useful *end_xml
All lines are optional for the flat file. 
Each line between *start_xml and *end_xml defines the output that will be
included in the out.csv files of the model. The applicable outputs can be found 
in the existing out.dictionary files. So, the process is (since not many people
outside nrcan are familiar with the xml outputs):
1) install esp-r with xml active
2) add these line in your flat file that describes the esp-r model
3) make sure you specify save level 5 in the flat file
4) generate esp-r model from the flat file (prj -file NameOfFlatFile -act
silentread)
5) having these xml lines (or others from the out.dictionary) in the flat file, 
it will generate the input.xml in the model's cfg folder (i.e. it will be generated 
together with the rest of the model).
6) Run a simulation. Your pre-specified results are in out.csv without having to
generate the results analysis.

This commit also updates the documentation files. Having however 2 files for
documentation that contain the same information is problematic and generates
mistakes. This commit covers gaps that were generated by not updating both files
in previous commits.

There is also a placeholder added in the flat file for ipv related items but
this is not currently used anywhere and it is optional in the flat file.





Former-commit-id: 1fbdadc1b142267850c74dbcf9b56d5a70dcb83e